fix: исправление замечаний от 22.05.26

+++ app/routers/projects/projects.py
@@ -1,9 +1,10 @@
 from typing import Annotated, Optional, Sequence
 from uuid import UUID
 
-from fastapi import APIRouter, Query
+from fastapi import APIRouter, Depends, HTTPException, Query, status
 
-from app.dependencies.auth import CurrentStudentDep
+from app.core.config import settings
+from app.dependencies.auth import CurrentStudentDep, MemberRole, ProjectRoleDep
 from app.dependencies.services import (
     ProjectMemberServiceDep,
     ProjectServiceDep,
@@ -30,35 +31,31 @@ router.include_router(projects_teams.router)
 router.include_router(projects_members.router)
 
 
-@router.get('/')
+@router.get('/', dependencies=[Depends(CurrentStudentDep)])
 async def get_projects(
-    _student: CurrentStudentDep,
     project_service: ProjectServiceDep,
     filters: Annotated[ProjectFilters, Query()]
 ) -> Sequence[ProjectPublic]:
     return await project_service.get_projects(filters)
 
 
-@router.post('/')
+@router.post('/', dependencies=[Depends(CurrentStudentDep)])
 async def create_project(
-    _student: CurrentStudentDep,
     project_service: ProjectServiceDep, project_create: ProjectCreate
 ) -> ProjectPublic:
     return await project_service.create_project(project_create)
 
 
-@router.post('/join')
+@router.post('/join', dependencies=[Depends(CurrentStudentDep)])
 async def join_project(
-    _student: CurrentStudentDep,
     project_member_service: ProjectMemberServiceDep,
     pm_create: ProjectMemberCreate
 ) -> Optional[ProjectMemberPublic]:
     return await project_member_service.add_member_to_project(pm_create)
 
 
-@router.get('/{project_id}')
+@router.get('/{project_id}', dependencies=[Depends(CurrentStudentDep)])
 async def get_project(
-    _student: CurrentStudentDep,
     project_service: ProjectServiceDep, project_id: UUID
 ) -> Optional[ProjectPublic]:
     return await project_service.get_project(project_id)
@@ -66,18 +63,32 @@ async def get_project(
 
 @router.put('/{project_id}')
 async def update_project(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     project_service: ProjectServiceDep,
     project_update: ProjectUpdate,
     project_id: UUID,
 ) -> Optional[ProjectPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await project_service.update_project(project_update, project_id)
 
 
 @router.delete('/{project_id}')
 async def detele_project(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     project_member_service: ProjectServiceDep,
     project_id: UUID,
 ) -> Optional[ProjectPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await project_member_service.delete_project(project_id)
