+++ app/routers/sprints/sprints.py
@@ -1,9 +1,10 @@
 from typing import Optional, Sequence
 from uuid import UUID
 
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException, status
 
-from app.dependencies.auth import CurrentStudentDep
+from app.core.config import settings
+from app.dependencies.auth import CurrentStudentDep, MemberRole, ProjectRoleDep
 from app.dependencies.services import SprintServiceDep
 from app.models.sprints.sprint import (
     SprintCreate,
@@ -25,50 +26,82 @@ router.include_router(project_tasks.router)
 
 @router.get('/')
 async def get_sprints(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     sprint_service: SprintServiceDep, project_id: UUID, filters: SprintFilters
 ) -> Sequence[SprintPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     filters.project_id = project_id
     return await sprint_service.get_sprints(filters)
 
 
 @router.post('/')
 async def create_sprint(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     sprint_service: SprintServiceDep,
     sprint_create: SprintCreate,
     project_id: UUID,
 ) -> SprintPublic:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     sprint_create.project_id = project_id
     return await sprint_service.create_sprint(sprint_create)
 
 
 @router.get('/{sprint_id}')
 async def get_sprint(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     sprint_service: SprintServiceDep,
-    _project_id: UUID,
     sprint_id: UUID,
 ) -> Optional[SprintPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role == MemberRole.OTHER
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await sprint_service.get_sprint(sprint_id)
 
 
 @router.put('/{sprint_id}')
 async def update_sprint(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     sprint_service: SprintServiceDep,
-    _project_id: UUID,
     sprint_id: UUID,
     sprint_update: SprintUpdate,
 ) -> Optional[SprintPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await sprint_service.update_sprint(sprint_update, sprint_id)
 
 
 @router.delete('/{sprint_id}')
 async def delete_sprint(
-    _student: CurrentStudentDep,
+    student: CurrentStudentDep,
+    project_role: ProjectRoleDep,
     sprint_service: SprintServiceDep,
-    _project_id: UUID,
     sprint_id: UUID,
 ) -> Optional[SprintPublic]:
+    if (
+        student.role == settings.role.default_user_role_code
+        and project_role != MemberRole.TEAMLEAD
+    ):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
     return await sprint_service.delete_sprint(sprint_id)
